fix(dropbox): Address code review comments.
