+++ lang/main.json
@@ -450,10 +450,11 @@
         "on": "Recording",
         "pending": "Preparing to record the meeting...",
         "rec": "REC",
-        "authDropboxText": "To start the recording you need to first authorize our Dropbox recording app. After the recording is finished the file will be uploaded to your Dropbox account.",
-        "authDropboxCompletedText": "Our Dropbox app has been authorized successfully. You should see the recorded file in your Dropbox account shortly after the recording has finished.",
+        "authDropboxText": "Upload your recording to Dropbox.",
+        "authDropboxCompletedText": "Your recording file will appear in your Dropbox shortly after the recording has finished.",
         "serviceName": "Recording service",
         "signOut": "Sign Out",
+        "signIn": "sign in",
         "loggedIn": "Logged in as __userName__",
         "availableSpace": "Available space: __spaceLeft__ MB (approximately __duration__ minutes of recording)",
         "startRecordingBody": "Are you sure you would like to start recording?",
