+++ react/features/dropbox/actions.js
@@ -2,11 +2,39 @@
 
 import { Dropbox } from 'dropbox';
 
-import { getLocationContextRoot, parseStandardURIString } from '../util';
-import { parseURLParams } from '../config';
+import {
+    getJitsiMeetGlobalNS,
+    getLocationContextRoot,
+    parseStandardURIString
+} from '../base/util';
+import { parseURLParams } from '../base/config';
 
-import { authorize } from './functions';
 import { UPDATE_DROPBOX_TOKEN } from './actionTypes';
+
+/**
+ * Executes the oauth flow.
+ *
+ * @param {string} authUrl - The URL to oauth service.
+ * @returns {Promise<string>} - The URL with the authorization details.
+ */
+function authorize(authUrl: string): Promise<string> {
+    const windowName = `oauth${Date.now()}`;
+    const gloabalNS = getJitsiMeetGlobalNS();
+
+    gloabalNS.oauthCallbacks = gloabalNS.oauthCallbacks || {};
+
+    return new Promise(resolve => {
+        const popup = window.open(authUrl, windowName);
+
+        gloabalNS.oauthCallbacks[windowName] = () => {
+            const returnURL = popup.location.href;
+
+            popup.close();
+            delete gloabalNS.oauthCallbacks.windowName;
+            resolve(returnURL);
+        };
+    });
+}
 
 /**
  * Action to authorize the Jitsi Recording app in dropbox.
