+++ react/features/dropbox/index.js
@@ -1,3 +1,4 @@
 export * from './actions';
+export * from './functions';
 
 import './reducer';
