+++ react/features/recording/functions.js
@@ -1,10 +1,6 @@
 // @flow
 
-import { Dropbox } from 'dropbox';
-
 import { JitsiRecordingConstants } from '../base/lib-jitsi-meet';
-
-const logger = require('jitsi-meet-logger').getLogger(__filename);
 
 /**
  * Searches in the passed in redux state for an active recording session of the
@@ -34,38 +30,4 @@ export function getActiveSession(state: Object, mode: string) {
 export function getSessionById(state: Object, id: string) {
     return state['features/recording'].sessionDatas.find(
         sessionData => sessionData.id === id);
-}
-
-/**
- * Fetches information about the user's dropbox account.
- *
- * @param {string} token - The dropbox access token.
- * @param {string} clientId - The Jitsi Recorder dropbox app ID.
- * @returns {Promise<Object|undefined>}
- */
-export function getDropboxData(
-        token: string,
-        clientId: string
-): Promise<?Object> {
-    const dropboxAPI = new Dropbox({
-        accessToken: token,
-        clientId
-    });
-
-    return Promise.all(
-        [ dropboxAPI.usersGetCurrentAccount(), dropboxAPI.usersGetSpaceUsage() ]
-    ).then(([ account, space ]) => {
-        const { allocation, used } = space;
-        const { allocated } = allocation;
-
-        return {
-            userName: account.name.display_name,
-            spaceLeft: Math.floor((allocated - used) / 1048576)// 1MiB=1048576B
-        };
-
-    }, error => {
-        logger.error(error);
-
-        return undefined;
-    });
 }
